pyproject.toml
[project]
name = "agentic-vs-vanilla-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "chromadb>=1.4.1",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.7",
    "langchain-text-splitters>=1.1.0",
    "lxml>=6.0.2",
    "openai>=2.17.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.3.2",
    "sentence-transformers>=5.2.2",
    "typer>=0.21.1",
]
